Create FoundStats_2023_PS3.html
## Part 4 Reflection (25 pts)
Answer the following questions in a couple of sentences each. No right or wrong answers! (5 pts each)

1. Since we have reached the end of the first half of the term, reflect on your coding skills. How does where you are now compare to how you felt at the beginning of this term?

I think have definitely have more practice with R which makes me more comfortable using it in new ways.Unix/Linux looked more intimidating in the beginning then it actually was and I believe I have a solid foundation in it now.

2. What aspects of the class structure have been helpful and that you'd like to see continued?

I like how we have the ability to practice code during class and that it's not all just lecture. I also like how the homework is directly related to what we learned in class that week.

3. Are there aspects of the class structure that could be improved to help your learning?

Perhaps talking more about the logic of how these functions work. What I mean is spending more time defining things like what a string is or what an index is can help us when during our homework and in class practice. Especially in the beginning, spending more time getting us immersed in what the vocabulary of coding is. I wish the class what a bit longer honestly (like an extra 20 minutes) so that we can discuss the more conceptual while maintaining the in class practice.

4. Take a look at the syllabus schedule for the second half of the term. Are there any topics in particular that seem more exciting or useful for you?

I feel most excited about the statistics portion of the class because I want to know how use R for this purpose. I think this will be the majority of what I'll use R for in my research.

5. Are there any other basic coding skills that you wish we would have covered in this class, that we could try to incorporate in this class for next year?

Perhaps learning about how to translate the skills from one language into another. I feel like many coding languages have different vocabulary for the same thing. So for example, comparing R with Python might be a good way to apply our knowledge of R to another programming language.